refactor(dashboard): page redirect to available tenant
