+++ dashboard/src/app/[tenantId]/(diagram)/components/NodeTypes/ExternalEvent/ExternalEvent.tsx
@@ -4,7 +4,6 @@ import { useQuery } from '@tanstack/react-query'
 import { formatTime, getVariableValue } from '@/app/utils'
 import { LHStatus, Node as NodeProto } from 'littlehorse-client/proto'
 import { ExternalLinkIcon, EyeIcon, MailOpenIcon } from 'lucide-react'
-import Link from 'next/link'
 import { FC, memo, useCallback } from 'react'
 import { Handle, Position } from 'reactflow'
 import { NodeProps } from '..'
