+++ dashboard/src/app/[tenantId]/(diagram)/components/NodeTypes/Task/TaskDetails.tsx
@@ -4,7 +4,6 @@ import { useWhoAmI } from '@/contexts/WhoAmIContext'
 import { useQuery } from '@tanstack/react-query'
 import { NodeRun, TaskNode } from 'littlehorse-client/proto'
 import { ExternalLinkIcon, EyeIcon } from 'lucide-react'
-import Link from 'next/link'
 import { FC, useCallback } from 'react'
 import { useModal } from '../../../hooks/useModal'
 import { NodeDetails } from '../NodeDetails'
