+++ dashboard/src/app/[tenantId]/(diagram)/wfRun/[...ids]/components/Details.tsx
@@ -2,7 +2,6 @@
 import LinkWithTenant from '@/app/[tenantId]/components/LinkWithTenant'
 import { concatWfRunIds, formatDate } from '@/app/utils'
 import { WfRun } from 'littlehorse-client/proto'
-import Link from 'next/link'
 import { FC } from 'react'
 
 type DetailsProps = WfRun
