+++ dashboard/src/app/[tenantId]/(diagram)/wfSpec/[...props]/components/SearchVariableDialog.tsx
@@ -15,7 +15,6 @@ import { WithBookmark } from '@/types'
 import { useInfiniteQuery } from '@tanstack/react-query'
 import { VariableDef, VariableValue, WfSpec } from 'littlehorse-client/proto'
 import { RefreshCwIcon } from 'lucide-react'
-import Link from 'next/link'
 import { FC, Fragment, useState } from 'react'
 import { useDebounce } from 'use-debounce'
 import { searchVariables } from '../actions/searchVariables'
