+++ dashboard/src/app/[tenantId]/(diagram)/wfSpec/[...props]/components/WfRuns.tsx
@@ -6,7 +6,6 @@ import { useWhoAmI } from '@/contexts/WhoAmIContext'
 import { useInfiniteQuery } from '@tanstack/react-query'
 import { WfSpec, lHStatusFromJSON } from 'littlehorse-client/proto'
 import { RefreshCwIcon } from 'lucide-react'
-import Link from 'next/link'
 import { useSearchParams } from 'next/navigation'
 import { FC, Fragment, useMemo, useState } from 'react'
 import { PaginatedWfRunIdList, searchWfRun } from '../actions/searchWfRun'
