+++ dashboard/src/app/[tenantId]/components/VersionSelector.tsx
@@ -1,6 +1,5 @@
 import { Listbox } from '@headlessui/react'
 import { TagIcon } from 'lucide-react'
-import Link from 'next/link'
 import { FC } from 'react'
 import LinkWithTenant from './LinkWithTenant'
