+++ dashboard/src/app/[tenantId]/components/tables/WfSpecTable.tsx
@@ -1,6 +1,5 @@
 import { WfSpecId } from 'littlehorse-client/proto'
 import { TagIcon } from 'lucide-react'
-import Link from 'next/link'
 import { FC, Fragment } from 'react'
 import { SearchResultProps } from '.'
 import { useParams } from 'next/navigation'
