+++ dashboard/src/app/[tenantId]/components/tables/WorkflowEventDefTable.tsx
@@ -1,5 +1,4 @@
 import { WorkflowEventDefId } from 'littlehorse-client/proto'
-import Link from 'next/link'
 import { FC, Fragment } from 'react'
 import { SearchResultProps } from '.'
 import { useParams } from 'next/navigation'
