+++ dashboard/src/app/[tenantId]/externalEventDef/[name]/components/ExternalEventDef.tsx
@@ -11,7 +11,6 @@ import { useWhoAmI } from '@/contexts/WhoAmIContext'
 import { useInfiniteQuery } from '@tanstack/react-query'
 import { ExternalEventDef as ExternalEventDefProto } from 'littlehorse-client/proto'
 import { RefreshCwIcon } from 'lucide-react'
-import Link from 'next/link'
 import { FC, Fragment, useState } from 'react'
 import { PaginatedExternalEventList, searchExternalEvent } from '../actions/searchExternalEvent'
 import { Details } from './Details'
