+++ dashboard/src/app/[tenantId]/userTaskDef/[...props]/components/UserTaskDef.tsx
@@ -15,7 +15,6 @@ import { Button, Field, Input, Label } from '@headlessui/react'
 import { useInfiniteQuery } from '@tanstack/react-query'
 import { UserTaskDef as UserTaskDefProto, UserTaskRunStatus } from 'littlehorse-client/proto'
 import { RefreshCwIcon } from 'lucide-react'
-import Link from 'next/link'
 import React, { FC, Fragment, useState } from 'react'
 import { useDebounce } from 'use-debounce'
 import { Details } from './Details'
