+++ dashboard/src/app/[tenantId]/workflowEventDef/[name]/components/WorkflowEventDef.tsx
@@ -10,7 +10,6 @@ import { useWhoAmI } from '@/contexts/WhoAmIContext'
 import { useInfiniteQuery } from '@tanstack/react-query'
 import { WorkflowEventDef as WorkflowEventDefProto } from 'littlehorse-client/proto'
 import { RefreshCwIcon } from 'lucide-react'
-import Link from 'next/link'
 import { FC, Fragment, useState } from 'react'
 import { PaginatedWorkflowEventList, searchWorkflowEvent } from '../actions/searchWorkflowEvent'
 import { Details } from './Details'
